Crate oak_app_build

An utility for building an Oak application from a manifest file.

Invoke with:

cargo run --manifest-path=sdk/rust/oak_app_build/Cargo.toml -- \
    --manifest-path=<OAK_APP_MANIFEST_FILE>

The compiled Oak application will then be stored under a bin subdirectory from the manifest file location, named after the name field from the app manifest, and with an .oak extension.

Structs

External

InitialNodeConfig

Manifest

Opt
Command line options for the Oak Application builder.

Enums

Module

Constants

CACHE_DIRECTORY
Directory used to save downloaded Wasm modules.

OUTPUT_DIRECTORY
Directory used to save build artifacts.

Functions

download_file_from_url
Download file from url.

get_module_cache_path
Get path for caching a downloaded file in CACHE_DIRECTORY. Cache file is named after sha256_sum.

get_output_path
Get path for the output application file in OUTPUT_DIRECTORY.

get_sha256
Computes SHA256 sum from data and returns it as a HEX encoded string.

load_module
Load Wasm module from file or URL if specified. If the file was downloaded from URL, it is cached in CACHE_DIRECTORY.

main

write_config_to_file
Serializes an application configuration from app_config and writes it into filename.